Rename: 컴포넌트 이름 오타 수정
