+++ src/app/admin/notice/create/_components/ContentsContainer.tsx
@@ -9,7 +9,7 @@ import BodyContent from './BodyContent';
 import SubTitleContent from './SubTitleContet';
 import ButtonContent from './ButtonContent';
 import ImageContent from './ImageContent';
-import LineContent from './LintContent';
+import LineContent from './LineContent';
 import NoteContent from './NoteContent';
 
 interface FormAboutContentProps {
